useTransactionListForPendingTxs to handle pending txs
